core/core/urls.py
"""
URL configuration for core project.

The `urlpatterns` list routes URLs to views. For more information please see:
    https://docs.djangoproject.com/en/4.2/topics/http/urls/
Examples:
Function views
    1. Add an import:  from my_app import views
    2. Add a URL to urlpatterns:  path('', views.home, name='home')
Class-based views
    1. Add an import:  from other_app.views import Home
    2. Add a URL to urlpatterns:  path('', Home.as_view(), name='home')
Including another URLconf
    1. Import the include() function: from django.urls import include, path
    2. Add a URL to urlpatterns:  path('blog/', include('blog.urls'))
"""
from django.contrib import admin
from django.urls import path , include 
from rest_framework.schemas import get_schema_view 
from rest_framework.documentation import include_docs_urls 
from drf_spectacular.views import SpectacularAPIView, SpectacularSwaggerView , SpectacularRedocView


urlpatterns = [
    path('admin/', admin.site.urls),
    path('api/', include('news.urls')),
    path('api/', include('home.urls')),
    # path('api-auth/',include('rest_framework.urls')),
    path('api/auth/', include('dj_rest_auth.urls')),
    path('schema/',get_schema_view(title="blog api")),
    # path('api/v1/rest-auth', include('rest_auth.urls')),
    
    path('schema/', SpectacularAPIView.as_view(), name='schema'),
    path('docs/', SpectacularSwaggerView.as_view(url_name='schema')),
    path('redocs/',SpectacularRedocView.as_view(),name="redocs")
    
]
