core/requirements.txt
asgiref==3.11.0
attrs==25.4.0
certifi==2025.11.12
charset-normalizer==3.4.4
config==0.5.1
coreapi==2.3.3
coreschema==0.0.4
dj-rest-auth==7.0.1
Django==4.2.26
django-allauth==65.13.1
django-rest-auth==0.9.5
djangorestframework==3.16.1
drf-spectacular==0.29.0
drf-spectacular-sidecar==2025.12.1
idna==3.11
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
pillow==11.3.0
PyYAML==6.0.3
referencing==0.36.2
requests==2.32.5
rpds-py==0.27.1
six==1.17.0
sqlparse==0.5.4
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
